Use stdin for arcanist instead of piping via shell
This makes things a lot less hacky and allows us to output the command
we're running but scrub the conduit token (if present).

The BEST part of about this is we're no longer doing hacky string
escaping to avoid the fact that we're constructing JSON strings that
have to be shell-escaped.

We also now check for the results of the return code of calling arcanist.