Refactor PhabricatorNotifier comment builder
Instead of having one heaping class that's responsible for a bunch of
business logic, breaking out the comment building and code coverage
comparison into a self-contained class that is easier to test and understand.

Future refactors will break down the Notifier even further, but this is
one atomic piece of work that can go in.

Additionally, fixed a bug in the TestUtils default coverage data, which
was off by 100x for cobertura results (since cobertura the plugin
returns them as floats from 0-1)