Fix regression in posting comments
Introduced in c9ca27b, I was using the "diff" ID instead of the "revision" ID
when trying to post a comment.

I'm not super happy about the separation of concerns here and am open to
alternate approaches / suggestions for improvement.

I thought about making DifferentialClient fetch the Differential on
instantiation and store a reference, then calling `getRevisionID(false)`
internally when it wants to `postComment`, but that felt like unnecessary
coupling, and I didn't like the side-effect of fetching on instantiation.