send Harbormaster uris to link back to jenkins
It would be useful if Differential could provide a link back to
Jenkins since engineers will probably want to look at the results of
failing builds.  This can be accomplished by uploading a uri
"artifact" for display.

Ideally we would tell Harbormaster the uri as soon as the build is
started, but I didn't want to complicated this plugin by adding even
more extension points.

See https://secure.phabricator.com/T8659 for details.