Use Junit Error Trace as Unit Result details (#170)
* Use Junit Error Trace as Unit Result details

Harbormaster accept `details` for Unit Results. It would be more
informative for test report to have stack trace when test is failing.

For pass tests, the stack trace is null, which will be omitted by
JSONObject.

* Integration style test for Conduit Client

The existing fake conduit client doesn’t verify the requests. In order
to verify the data `sent` to conduit server, it `requestBodies` to
record every single request form body. The tests can use the recorded
request form body to verify if it is expected.