Show more info to the user when build failed due to the coverage drop
When code coverage drop exceeds some configured limit - the plugin will
mark the build as failed, but it's hard to figure out the reason of the
failure. To address this issue added logging of this event and also
sending of the reason of the failure as part of the comment posted to
phabricator.