Add minimum coverage threshold (#220)
* Add minimum coverage threshold

Right now delta of code coverage is analyzed always then code coverage
check is enabled. This makes this feature a bit annoying when you have
decent code coverage of the project in general and decrease line count
during refactoring of the code. Added minimum coverage threshold to
be able to disable the check for coverage drop when overall code
coverage is higher than this threshold. If coverage is lower than
minimum threshold - than code coverage delta will be analyzed.

* Fix checkstyle violations